% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{hessian}
\alias{hessian}
\title{Compute the second derivatives of the dependent variables with respect to the independent variable}
\usage{
hessian(s_arg1, s_arg2, s_arg3, s_arg4, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{n}{integer number of independents and m = 1}

\item{x}{independent vector of length n}

\item{H}{resulting Hessian matrix (n*n) \nabla^2F(x)}
}
\value{
None. the output is the the parameter H
}
\description{
\code{hessian} computes the Hessain matrix values for m dependents with respect to
n independents for the computation that has been traced.
}
\examples{
hessian(tag,n,x,H)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{autodiffadolc}}
