% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{trace_on}
\alias{trace_on}
\title{Start of the ADOL-C taping process}
\usage{
trace_on(...)
}
\arguments{
\item{tag}{integer to identify the trace that is created}
}
\value{
None
}
\description{
\code{trace_on} starts the \code{autodiffadolc} tracing process. Using 
\code{autodiffadolc} consists of two phases (1) Tracing (also called taping), 
generally performed once and (2) Computation of derivatives, which can be performed. 
\code{trace_on} is the initialization for the taping process. Creates buffers 
for this tape, sets files names, and calls appropriate setup routines.
}
\examples{
trace_on(1)

\\seealso{\\code{\\link{trace_off}}, \\code{\\link{badouble_declareIndependent}}, 
        \\code{\\link{badouble_declareDependent}}}
        
\\keyword{autodiffadolc}

}
